A version of RPI.SenseHat that works with Lightning
This version of the RPI.SenseHat works with either the default drivers
or the Lighting (DMAP) drivers on IOT Core.
The change adds a dependency on the latest Lightning Nuget package,
v1.1.0-alpha and updates the sources, so it default to the Lightning
driver if enabled.
Tested on RPI2 using IOT Core preview build, 14342
(https://www.microsoft.com/en-us/software-download/windowsiot)
Lightning required upgrading the solutions to Windows SDK 10586.